+++ ecms_acquia/tests/src/Functional/InstallationTest.php
@@ -32,9 +32,25 @@ class InstallationTest extends AllProfileInstallationTestsAbstract {
   protected $defaultTheme = 'stark';
 
   /**
+   * Run all available tests.
+   *
+   * This combines all of the functional tests into one allowing for only one
+   * Drupal installation. This should significantly increase the speed of
+   * all of the tests.
+   */
+  public function testAllTheThings(): void {
+    // Run our profile tests only.
+    $this->ensureAcsfModulesEnabled();
+
+    // Run all of the global tests.
+    $this->globalTests();
+
+  }
+
+  /**
    * Test that the ACSF modules are installed.
    */
-  public function testAcsfModulesEnabled(): void {
+  private function ensureAcsfModulesEnabled(): void {
     $account = $this->drupalCreateUser(['administer modules']);
     $this->drupalLogin($account);
 
@@ -45,6 +61,7 @@ class InstallationTest extends AllProfileInstallationTestsAbstract {
     $this->assertSession()->checkboxChecked('edit-modules-acsf-duplication-enable');
     $this->assertSession()->checkboxChecked('edit-modules-acsf-theme-enable');
     $this->assertSession()->checkboxChecked('edit-modules-acsf-variables-enable');
+    $this->drupalLogout($account);
   }
 
 }